pyproject.toml
[project]
name = "speechlanguageidapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.0",
    "librosa>=0.11.0",
    "numpy>=2.3.5",
    "onnxruntime>=1.23.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
